% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blast_results_to_featuredata.R
\name{add_blast_results_to_featuredata}
\alias{add_blast_results_to_featuredata}
\title{add_interpro_to_featuredata}
\usage{
add_blast_results_to_featuredata(opt = opt, blastanalyses = NULL)
}
\description{
JAMSalpha function
}
