% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_enterotypes_to_metadata.R
\name{add_enterotypes_to_metadata}
\alias{add_enterotypes_to_metadata}
\title{add_enterotypes_to_metadata(ptable = NULL, ExpObj = NULL, PPM_normalize_to_bases_sequenced = TRUE, force_n_clusters = NULL)
Defines enterotypes and attributes samples to enterotypes}
\usage{
add_enterotypes_to_metadata(
  ptable = NULL,
  ExpObj = NULL,
  PPM_normalize_to_bases_sequenced = TRUE,
  force_n_clusters = NULL
)
}
\description{
add_enterotypes_to_metadata(ptable = NULL, ExpObj = NULL, PPM_normalize_to_bases_sequenced = TRUE, force_n_clusters = NULL)
Defines enterotypes and attributes samples to enterotypes
}
