% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_interpro_to_featuredata.R
\name{add_interpro_to_featuredata}
\alias{add_interpro_to_featuredata}
\title{add_interpro_to_featuredata}
\usage{
add_interpro_to_featuredata(opt = NULL, doinparallel = FALSE)
}
\description{
JAMSalpha function
}
