% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{add_shape_to_plot_safely}
\alias{add_shape_to_plot_safely}
\title{add_shape_to_plot_safely(p = NULL, shapevec = NULL, cdict = NULL)}
\usage{
add_shape_to_plot_safely(
  p = NULL,
  shapevec = NULL,
  shapeby = NULL,
  cdict = NULL
)
}
\description{
#Given a vector of classes for adding shapes to a ggplot, attributes shapes safely in the presence or absence of a cdict containing shape info
}
