% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agglomerate_features.R
\name{agglomerate_features}
\alias{agglomerate_features}
\title{agglomerate_features(mgseqobj=NULL, glomby=NULL)}
\usage{
agglomerate_features(mgseqobj = NULL, glomby = NULL)
}
\description{
Agglomerates features in a JAMS mgSeq object safely
}
