% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{analyze_fastq_filename}
\alias{analyze_fastq_filename}
\title{analyze_fastq_filename(fn = NULL, maintain_Illumina_format = FALSE)}
\usage{
analyze_fastq_filename(fn = NULL, maintain_Illumina_format = FALSE)
}
\description{
Given a fastq filename, will return a list with facts about its anatomy.
}
