% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_databases.R
\name{blast_databases}
\alias{blast_databases}
\title{blast_databases}
\usage{
blast_databases(opt = NULL, blastanalyses = NULL, QcovThreshold = 75,
  HitcovThreshold = 75, PidentThreshold = 75)
}
\description{
JAMSalpha function
}
