% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contig_depth_by_length.R
\name{calculate_contig_depth_by_length}
\alias{calculate_contig_depth_by_length}
\title{calculate_contig_depth_by_length}
\usage{
calculate_contig_depth_by_length(
  contigsdata = opt$contigsdata,
  featuredata = opt$featuredata,
  max_length_thresh = 2000
)
}
\description{
JAMSalpha function
}
