% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{calculate_safe_threads}
\alias{calculate_safe_threads}
\title{Calculate safe number of threads to avoid Out-Of-Memory errors when using contextualize_taxonomy}
\usage{
calculate_safe_threads(
  requested_threads = NULL,
  totmembytes = NULL,
  list.data = NULL,
  BinsDF = NULL
)
}
\arguments{
\item{requested_threads}{The maximum number of CPUs requested by the user.}

\item{totmembytes}{The total available system memory in bytes.}

\item{list.data}{The list.data object that will be exported to workers.}

\item{BinsDF}{The BinsDF object that will be exported to workers.}
}
\value{
An integer representing the safe number of threads to use.
}
\description{
Calculate safe number of threads to avoid Out-Of-Memory errors when using contextualize_taxonomy
}
