% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{can_be_made_numeric}
\alias{can_be_made_numeric}
\title{can_be_made_numeric(x, cats_to_ignore = NULL)
Returns logical for if a vector can be coerced into numeric.}
\usage{
can_be_made_numeric(x, cats_to_ignore = NULL)
}
\description{
can_be_made_numeric(x, cats_to_ignore = NULL)
Returns logical for if a vector can be coerced into numeric.
}
