% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{check_for_banked_samples}
\alias{check_for_banked_samples}
\title{check_for_banked_samples(JAMSarchive_path = NULL)}
\usage{
check_for_banked_samples(JAMSarchive_path = NULL)
}
\description{
Given an absolute path to a JAMSarchive style directory (wherein jamsfiles and JAMStarballs are), this function will return a vector containing the names of samples for which there are both a .jams file and a JAMS alpha run tarball. This may be useful to quickly ascertiain which samples have been banked or not, in the scope of project management.
}
