% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_16S_from_contigs.R
\name{classify_16S_from_contigs}
\alias{classify_16S_from_contigs}
\title{classify_16S_from_contigs(opt=opt)}
\usage{
classify_16S_from_contigs(opt = NULL)
}
\description{
Classifies taxonomically 16S features found in contig annotations using Dada2
}
