% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ucobias.R
\name{compare_ucobias}
\alias{compare_ucobias}
\title{compare_ucobias}
\usage{
compare_ucobias(featuredata = NULL, contigsdata = NULL,
  taxlevel = "LKT", taxaofinterest = NULL, ucobias = NULL,
  method = "tSNE", frac = 0.25)
}
\description{
Compares taxa by codon usage bias.
}
