% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_metadata.R
\name{compose_metadata}
\alias{compose_metadata}
\title{compose_metadata}
\usage{
compose_metadata(metadatafolder = NULL, recursive = FALSE,
  prefixsubstitutiontable = NULL, restraintoprefixes = NULL)
}
\description{
Returns JAMS-style metadata for a vector of prefixes based on data available in a metadata farm (folder with one or several metadata files)
This function is experimental. Use with caution.
}
