% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_signature_numbases.R
\name{compute_signature_numbases}
\alias{compute_signature_numbases}
\title{compute_signature_numbases}
\usage{
compute_signature_numbases(featuredata = NULL, columnname = NULL)
}
\description{
JAMSalpha function
}
