% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{compute_x_font}
\alias{compute_x_font}
\title{compute_x_font(nx = 1, upper_n = 400, upper_fs = 0.1, lower_n = 10, lower_fs = 8, cex = 0.7)}
\usage{
compute_x_font(
  nx = 1,
  upper_n = 400,
  upper_fs = 0.1,
  lower_n = 10,
  lower_fs = 8,
  cex = 0.7
)
}
\description{
Compute the appropriate size of font to be used in JAMS heatmap column (sample) annotations.
}
