% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_entities_in_sample.R
\name{consolidate_entities_in_sample}
\alias{consolidate_entities_in_sample}
\title{consolidate_entities_in_sample(opt=opt)}
\usage{
consolidate_entities_in_sample(opt = opt)
}
\description{
JAMSalpha function
}
