% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_tables.R
\name{consolidate_tables}
\alias{consolidate_tables}
\title{consolidate_tables(opt=opt)}
\usage{
consolidate_tables(
  opt = opt,
  elements = NULL,
  taxonomic_spaces = c("Contig_LKT", "MB2bin", "ConsolidatedGenomeBin")
)
}
\description{
JAMSalpha function
}
