% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_JAMS_SEobj_to_wanted_taxa.R
\name{constrain_JAMS_SEobj_to_wanted_taxa}
\alias{constrain_JAMS_SEobj_to_wanted_taxa}
\title{Constrains a functional (non-taxonomic) JAMS-style SummarizedExperiment object which has feature-by-taxon stratification information (see JAMSbeta -u option) into a SummarizedExperiment object with BaseCounts and GeneCounts only pertaining to the wanted LKTs.}
\usage{
constrain_JAMS_SEobj_to_wanted_taxa(SEobj = NULL, wanted_LKTs = NULL)
}
\arguments{
\item{SEobj}{a functional (non-taxonomic) JAMS-style SummarizedExperiment object which has feature-by-taxon stratification information (see JAMSbeta -u option)}

\item{wanted_LKTs}{vector of LKTs to constrain the functional features of the original SEobj into}
}
\description{
Constrains a functional (non-taxonomic) JAMS-style SummarizedExperiment object which has feature-by-taxon stratification information (see JAMSbeta -u option) into a SummarizedExperiment object with BaseCounts and GeneCounts only pertaining to the wanted LKTs.
}
