% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualize_taxonomy.R
\name{contextualize_taxonomy}
\alias{contextualize_taxonomy}
\title{contextualize_taxonomy(LKTdosesall = LKTdosesall, list.data = list.data, normalize_length = FALSE, dissimilarity_cutoff = 0.15, threads = 1, totmembytes = NULL)}
\usage{
contextualize_taxonomy(
  LKTdosesall = LKTdosesall,
  list.data = list.data,
  normalize_length = FALSE,
  dissimilarity_cutoff = 0.15,
  threads = 1,
  totmembytes = NULL
)
}
\description{
This is an internal function used exclusively within the make_SummarizedExperiments function to cluster taxonomic entities belonging to the same species taxid into functional clades. Do not attempt to use this out of this context.
}
