% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{convert_matrix_log2}
\alias{convert_matrix_log2}
\title{convert_matrix_log2(mat = NULL, transformation = NULL)}
\usage{
convert_matrix_log2(mat = NULL, transformation = NULL)
}
\description{
Wrapper for transforming a matrix to and from log2 formats
}
