% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{convert_matrix_PA}
\alias{convert_matrix_PA}
\title{convert_matrix_PA(mat = NULL, threshPA = 0)}
\usage{
convert_matrix_PA(mat = NULL, threshPA = 0)
}
\description{
Wrapper for transforming a matrix to and from log2 formats
}
