% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_handling_functions.R
\name{countfastq}
\alias{countfastq}
\title{countfastq(fastqfile)
Wrapper for counting number of reads and number of bases of a fastq file in the system.
Returns a vector with read counts and base counts.}
\usage{
countfastq(fastqfile)
}
\description{
countfastq(fastqfile)
Wrapper for counting number of reads and number of bases of a fastq file in the system.
Returns a vector with read counts and base counts.
}
