% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_contaminated_isolates.R
\name{deconvolute_contaminated_isolates}
\alias{deconvolute_contaminated_isolates}
\title{deconvolute_contaminated_isolates
This function is experimental.
Function to deconvolute contaminated isolates.}
\usage{
deconvolute_contaminated_isolates(
  opt = NULL,
  contigsdata = NULL,
  assemblystats = NULL,
  outputdir = NULL
)
}
\description{
JAMSalpha function
}
