% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{detectHardwareResources}
\alias{detectHardwareResources}
\title{detectHardwareResources()
Generic function for getting number of CPUs and total memory on Biowulf or otherwise
Returns a named vector with available CPUs and Memory in bytes}
\usage{
detectHardwareResources()
}
\description{
detectHardwareResources()
Generic function for getting number of CPUs and total memory on Biowulf or otherwise
Returns a named vector with available CPUs and Memory in bytes
}
