% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{dichotomize_variable_in_md}
\alias{dichotomize_variable_in_md}
\title{dichotomize_variable_in_md(md = NULL, column_to_dichotomize = NULL, threshold = NULL, func_for_threshold = "median")}
\usage{
dichotomize_variable_in_md(
  md = NULL,
  column_to_dichotomize = NULL,
  threshold = NULL,
  func_for_threshold = "median"
)
}
\description{
Translates a continuous variable in metadata into a dichotomous discrete variable, either by set threshold, or by being above or below a particular value computeby by func_for_threshold, e.g. median.
}
