% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SRA_reads.R
\name{download_SRA_reads}
\alias{download_SRA_reads}
\title{download_SRA_reads(SRAaccessions = NULL, outfolder = NULL, tempfolder = NULL, prefetch = TRUE, threads = 8)}
\usage{
download_SRA_reads(
  SRAaccessions = NULL,
  outfolder = NULL,
  tempfolder = NULL,
  prefetch = TRUE,
  threads = 8
)
}
\description{
Wrapper for safely downloading SRA reads within the JAMS framework. You will need to have SRAtoolkit installed on your system.
}
