% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_genome_completeness.R
\name{estimate_genome_completeness}
\alias{estimate_genome_completeness}
\title{estimate_genome_completeness(taxonomic_completeness = NULL, taxon = NULL)}
\usage{
estimate_genome_completeness(taxonomic_completeness = NULL, taxon = NULL)
}
\description{
JAMSalpha function. Given a JAMSalpha style taxonomic_completeness data frame, will estimate the genome completeness using the "Percent contig sum over expected genome size" method, for entities which cannot be evaluated using checkM (non-bacterial contigs or length too short.)
}
