% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_assembly.R
\name{evaluate_assembly}
\alias{evaluate_assembly}
\title{evaluate_assembly(opt=NULL)}
\usage{
evaluate_assembly(opt = NULL)
}
\description{
Calculates the genome completeness and assembly stats for each taxonomic level in a contigsdata dataframe.
}
