% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_consolidated_Quality_split_df.R
\name{evaluate_consolidated_Quality_split_df}
\alias{evaluate_consolidated_Quality_split_df}
\title{evaluate_consolidated_Quality_split_df(consolidated_df = NULL)}
\usage{
evaluate_consolidated_Quality_split_df(consolidated_df = NULL)
}
\description{
JAMSalpha function
}
