% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_LKTs.R
\name{evaluate_LKTs}
\alias{evaluate_LKTs}
\title{evaluate_LKTs(opt = opt, contigsdata = "contigsdata", output_list_name = "taxonomic_completeness_and_counts")}
\usage{
evaluate_LKTs(
  opt = opt,
  contigsdata_name = "contigsdata",
  output_list_name = "taxonomic_completeness_and_counts",
  taxlvls = c("LKT", "Species")
)
}
\description{
JAMSalpha function
}
