% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_metaBAT_bin.R
\name{evaluate_metaBAT_bin}
\alias{evaluate_metaBAT_bin}
\title{evaluate_metaBAT_bin(opt=NULL)}
\usage{
evaluate_metaBAT_bin(opt = NULL)
}
\description{
Calculates the genome completeness and assembly stats for metaBAT bins.
}
