% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_metaBAT_bin_CheckM.R
\name{evaluate_metaBAT_bin_CheckM}
\alias{evaluate_metaBAT_bin_CheckM}
\title{evaluate_metaBAT_bin_CheckM(opt=NULL)}
\usage{
evaluate_metaBAT_bin_CheckM(opt = NULL, verbose = FALSE)
}
\description{
Calculates the genome completeness and assembly stats for metaBAT bins using CheckM (must be installed).
}
