% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{exporttabletsv}
\alias{exporttabletsv}
\title{exporttabletsv(dataobj = NULL, project = NULL, basename = NULL, row.names = TRUE, col.names = TRUE, path = NULL)
Generic function to export a dataframe to tsv format}
\usage{
exporttabletsv(
  dataobj = NULL,
  project = NULL,
  basename = NULL,
  row.names = TRUE,
  col.names = TRUE,
  path = NULL
)
}
\description{
exporttabletsv(dataobj = NULL, project = NULL, basename = NULL, row.names = TRUE, col.names = TRUE, path = NULL)
Generic function to export a dataframe to tsv format
}
