% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{extract_NCBI_taxid_from_featname}
\alias{extract_NCBI_taxid_from_featname}
\title{extract_NCBI_taxid_from_featname(Taxon = NULL, NCBI_taxonomic_rank = NULL)}
\usage{
extract_NCBI_taxid_from_featname(Taxon = NULL)
}
\description{
Securely extracts the NCBI taxid from a JAMS2-style taxonomic feature name
}
