% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sample_by_class_to_ignore.R
\name{filter_sample_by_class_to_ignore}
\alias{filter_sample_by_class_to_ignore}
\title{filter_sample_by_class_to_ignore(SEobj = NULL, mgseqobj = NULL, variables = NULL, class_to_ignore = NULL)}
\usage{
filter_sample_by_class_to_ignore(
  SEobj = NULL,
  mgseqobj = NULL,
  variables = NULL,
  class_to_ignore = NULL
)
}
\description{
Filters a SummarizedExperiment or metagenomeSeq object by several criteria.
}
