% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequence_by_length.R
\name{filter_sequence_by_length}
\alias{filter_sequence_by_length}
\title{filter_sequence_by_length(sequence=NULL, minlength=0, maxlength=Inf)}
\usage{
filter_sequence_by_length(sequence = NULL, minlength = 0,
  maxlength = Inf)
}
\description{
Filters sequences by size
}
