% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sequence_by_name.R
\name{filter_sequence_by_name}
\alias{filter_sequence_by_name}
\title{filter_sequence_by_name(input_sequences=NULL, sequencenames=NULL, keep=TRUE)}
\usage{
filter_sequence_by_name(
  input_sequences = NULL,
  sequencenames = NULL,
  keep = TRUE
)
}
\description{
Filters sequinr sequences by name either keeping or discarding specified sequence names
}
