% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{find16SrRNA}
\alias{find16SrRNA}
\title{find16SrRNA(contigs2length = NULL, opt = NULL)
JAMSalpha function to find number of 16S genes from contigs2length}
\usage{
find16SrRNA(contigs2length = NULL, opt = NULL)
}
\description{
find16SrRNA(contigs2length = NULL, opt = NULL)
JAMSalpha function to find number of 16S genes from contigs2length
}
