% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_accession_from_keyterm.R
\name{find_accession_from_keyterm}
\alias{find_accession_from_keyterm}
\title{find_accession_from_keyterm(expvec = NULL, analyses = NULL, SEobj = NULL, keyterms = NULL)}
\usage{
find_accession_from_keyterm(
  expvec = NULL,
  analyses = NULL,
  SEobj = NULL,
  keyterms = NULL
)
}
\description{
Returns the accession number(s) of a feature(s) matching a general expression present in a SummarizedExperiment or metagenomeSeq object.
}
