% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{find_clusters_in_matrix}
\alias{find_clusters_in_matrix}
\title{find_clusters_in_matrix(input_matrix = NULL, kmeans_split = NULL)}
\usage{
find_clusters_in_matrix(input_matrix = NULL, kmeans_split = NULL)
}
\description{
#Splits a matrix with counts in to k clusters by k-means clustering
}
