% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_kraken.R
\name{fix_kraken}
\alias{fix_kraken}
\title{fix_kraken(taxtable)}
\usage{
fix_kraken(taxtable = NULL, keepstrain = FALSE)
}
\description{
Fixes the mpa style output of a kaken.labels file. To be used only by the YAMSintegrate.R script.
}
