% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_qiime_taxonomy.R
\name{fix_qiime2_taxonomy}
\alias{fix_qiime2_taxonomy}
\title{Convert qiime2 taxonomy table to one compatible with JAMS}
\usage{
fix_qiime2_taxonomy(taxtsv)
}
\arguments{
\item{taxtsv}{a "taxonomy.tsv" file created by qiime2.}
}
\description{
Convert qiime2 taxonomy table to one compatible with JAMS
}
\examples{
fix_qiime2_taxonomy("taxonomy.tsv")
}
