% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{generate_filename}
\alias{generate_filename}
\title{generate_filename(title = "ABC_look_at_me", add_date = TRUE, suffix = "pdf")}
\usage{
generate_filename(title = "ABC_look_at_me", add_date = TRUE, suffix = "pdf")
}
\description{
Wrapper for generating file names.
}
