% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{get_assembly_stats_by_taxlevel}
\alias{get_assembly_stats_by_taxlevel}
\title{get_assembly_stats_by_taxlevel(opt = NULL, taxlevel = NULL)
JAMSalpha function to compute assembly statistics by taxonomic level}
\usage{
get_assembly_stats_by_taxlevel(opt = NULL, taxlevel = NULL)
}
\description{
get_assembly_stats_by_taxlevel(opt = NULL, taxlevel = NULL)
JAMSalpha function to compute assembly statistics by taxonomic level
}
