% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{get_assembly_stats_by_taxon}
\alias{get_assembly_stats_by_taxon}
\title{get_assembly_stats_by_taxon(opt = NULL, taxlevel = NULL, taxinterest = NULL)
JAMSalpha function to compute assembly statistics by taxon}
\usage{
get_assembly_stats_by_taxon(opt = NULL, taxlevel = NULL, taxinterest = NULL)
}
\description{
get_assembly_stats_by_taxon(opt = NULL, taxlevel = NULL, taxinterest = NULL)
JAMSalpha function to compute assembly statistics by taxon
}
