% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_stats.R
\name{get_feature_stats}
\alias{get_feature_stats}
\title{get_feature_stats}
\usage{
get_feature_stats(
  opt = NULL,
  ucoindex = "freq",
  doContigs = TRUE,
  doFeatures = FALSE
)
}
\description{
JAMSalpha function
}
