% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_to_accession_table_parallel.R
\name{get_feature_to_accession_table_parallel}
\alias{get_feature_to_accession_table_parallel}
\title{get_feature_to_accession_table_parallel}
\usage{
get_feature_to_accession_table_parallel(opt = NULL)
}
\description{
JAMSalpha function
}
