% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting_functions.R
\name{get_feature_to_blast_result_table}
\alias{get_feature_to_blast_result_table}
\title{get_feature_to_blast_result_table}
\usage{
get_feature_to_blast_result_table(opt = NULL, blastanalysis = NULL)
}
\description{
JAMSalpha function
}
