% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_completeness.R
\name{get_genome_completeness}
\alias{get_genome_completeness}
\title{get_genome_completeness<-function(pheno=pheno, list.data=NULL))}
\usage{
get_genome_completeness(pheno = pheno, list.data = NULL)
}
\description{
Returns a data frame with the estimated number of complete genomes of given taxa in each sample
}
