% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{getcontigs2length}
\alias{getcontigs2length}
\title{getcontigs2length(opt = NULL, taxlevel = NULL, taxinterest = NULL)
JAMSalpha function to compute total contig length of a taxon from contigs2length}
\usage{
getcontigs2length(opt = NULL, taxlevel = NULL, taxinterest = NULL)
}
\description{
getcontigs2length(opt = NULL, taxlevel = NULL, taxinterest = NULL)
JAMSalpha function to compute total contig length of a taxon from contigs2length
}
