% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{getL50}
\alias{getL50}
\title{getL50(contigs2length)
JAMSalpha function to get L50 from contigs2length}
\usage{
getL50(contigs2length)
}
\description{
getL50(contigs2length)
JAMSalpha function to get L50 from contigs2length
}
