% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{getN50}
\alias{getN50}
\title{getN50(contigs2length)
JAMSalpha function to get N50 from contigs2length}
\usage{
getN50(contigs2length)
}
\description{
getN50(contigs2length)
JAMSalpha function to get N50 from contigs2length
}
