% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{getN90}
\alias{getN90}
\title{getN90(contigs2length)
JAMSalpha function to get N90 from contigs2length}
\usage{
getN90(contigs2length)
}
\description{
getN90(contigs2length)
JAMSalpha function to get N90 from contigs2length
}
