% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{getOrdinalNumber1}
\alias{getOrdinalNumber1}
\title{Author: Phillip Burger
Date: Sept 3, 2014
Purpose: Get the ordinal rank of a number.
http://www.phillipburger.net/wordpress/ordinal-number-suffix-function-in-r}
\usage{
getOrdinalNumber1(num)
}
\description{
Author: Phillip Burger
Date: Sept 3, 2014
Purpose: Get the ordinal rank of a number.
http://www.phillipburger.net/wordpress/ordinal-number-suffix-function-in-r
}
