% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glom_completeness.R
\name{glom_completeness}
\alias{glom_completeness}
\title{glom_completeness(completeness_df = NULL, taxlvl = NULL)}
\usage{
glom_completeness(
  completeness_df = NULL,
  taxlvl = NULL,
  HQ_completeness_threshold = 90,
  HQ_contamination_threshold = 5,
  MHQ_completeness_threshold = 70,
  MHQ_contamination_threshold = 10,
  MQ_completeness_threshold = 50,
  MQ_contamination_threshold = 10,
  LQ_completeness_threshold = 0,
  LQ_contamination_threshold = 10,
  High_contamination_threshold = 20
)
}
\description{
JAMSalpha function
}
