% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{hm_fontsize_computer}
\alias{hm_fontsize_computer}
\title{hm_fontsize_computer(mat_rownames = NULL, mat_colnames = NULL, upper_n = 400, upper_fs = 0.1, lower_n = 10, lower_fs = 8, cex = 0.7)}
\usage{
hm_fontsize_computer(
  mat_rownames = NULL,
  mat_colnames = NULL,
  upper_n = 400,
  upper_fs = 0.1,
  lower_n = 10,
  lower_fs = 8,
  cex = 0.7
)
}
\description{
Returns a vector of length 2 with values for the appropriate sizes of fonts to be used in JAMS heatmap columns (samples) and rows (features), respectively.
}
