% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{infer_column_type}
\alias{infer_column_type}
\title{infer_column_type(phenotable = NULL, colm = NULL, class_to_ignore = "N_A")}
\usage{
infer_column_type(phenotable = NULL, colm = NULL, class_to_ignore = "N_A")
}
\description{
From a phenotable, infers what kind of a column it is, discrete or continuous.
}
