% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_LKT.R
\name{infer_LKT}
\alias{infer_LKT}
\title{infer_LKT}
\usage{
infer_LKT(taxtable)
}
\description{
Infers the Last Known Taxon (LKT) from a taxonomic table or dataframe containing taxonomic information and adds an LKT column to the dataframe.
}
