% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_LKT_pr2.R
\name{infer_LKT_pr2}
\alias{infer_LKT_pr2}
\title{infer_LKT_pr2}
\usage{
infer_LKT_pr2(taxtable)
}
\description{
Infers the Last Known Taxon (LKT) from a taxonomic table or dataframe containing taxonomic information and adds an LKT column to the dataframe.
}
