% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{IO_jams_workspace_image}
\alias{IO_jams_workspace_image}
\title{IO_jams_workspace_image(opt = NULL, workspaceimage = NULL, threads = 8, operation = c("save", "load"))
Safe way of either loading or saving an R workspace image. If argument workspaceimage is null, workspace image file will be searched for in opt (opt$projimage). If that is also NULL, saving or loading is aborted. If the fastSave package () is installed, multi-threaded loading or saving will be used. If opt is passed, number of CPUs will be set to opt$threads, trumping the threads argument.}
\usage{
IO_jams_workspace_image(
  opt = NULL,
  workspaceimage = NULL,
  threads = 8,
  operation = NULL,
  verbose = FALSE
)
}
\description{
IO_jams_workspace_image(opt = NULL, workspaceimage = NULL, threads = 8, operation = c("save", "load"))
Safe way of either loading or saving an R workspace image. If argument workspaceimage is null, workspace image file will be searched for in opt (opt$projimage). If that is also NULL, saving or loading is aborted. If the fastSave package () is installed, multi-threaded loading or saving will be used. If opt is passed, number of CPUs will be set to opt$threads, trumping the threads argument.
}
