% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{is.redundant}
\alias{is.redundant}
\title{is.redundant(vec)}
\usage{
is.redundant(vec)
}
\description{
Returns logical for a vector being redundant, i.e. has at least one element repeated in it.
}
