% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_classify_taxonomy.R
\name{kraken_classify_taxonomy}
\alias{kraken_classify_taxonomy}
\title{kraken_classify_taxonomy}
\usage{
kraken_classify_taxonomy(opt = NULL, fastafile = NULL, confidence = 1e-05)
}
\description{
JAMSalpha function
}
