% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metadata_from_xl.R
\name{load_metadata_from_xl}
\alias{load_metadata_from_xl}
\title{load_metadata_from_xl}
\usage{
load_metadata_from_xl(xlsxFile = NULL)
}
\description{
Loads an Excel(TM) file and returns a vector containing 1) the metadata itself and 2) description of label types for the metadata, provided the Excel file contains these data on sheets 1 and 2, respsctively.
}
