% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_qiime2_data.R
\name{load_qiime2_data}
\alias{load_qiime2_data}
\title{load qiime2 files into metagenomeSeq object}
\usage{
load_qiime2_data(otu_table, tax_table, metadata_table, total = 1e+05)
}
\arguments{
\item{otu_table}{a relative abundance otu file created by qiime2}

\item{tax_table}{a taxonomy file created by qiime2}

\item{metadata_table}{a metatable file}

\item{total}{total number of desired reads per sample}
}
\description{
load qiime2 files into metagenomeSeq object
}
\examples{
load_qiime2_data("otu.tsv", "taxonomy.tsv", "metadata.tsv")
}
