% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadobj.R
\name{loadobj}
\alias{loadobj}
\title{loadobj(objfn = NULL, never_used_columns = c("AntiFam", "CDD", "Coils", "FunFam", "Gene3D", "PANTHER", "Phobius", "ProSitePatterns", "SFLD", "SignalP_EUK", "SignalP_GRAM_NEGATIVE", "SignalP_GRAM_POSITIVE", "SMART", "TIGRFAM", "MetaCyc")
JAMSbeta helper function used by load_jamsfiles_from_system.}
\usage{
loadobj(
  objfn = NULL,
  never_used_columns = c("AntiFam", "CDD", "Coils", "FunFam", "Gene3D", "PANTHER",
    "Phobius", "ProSitePatterns", "SFLD", "SignalP_EUK", "SignalP_GRAM_NEGATIVE",
    "SignalP_GRAM_POSITIVE", "SMART", "TIGRFAM", "MetaCyc")
)
}
\description{
Do not fiddle with this.
}
