% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{log2PPMtoPct}
\alias{log2PPMtoPct}
\title{log2PPMtoPct(log2PPM)}
\usage{
log2PPMtoPct(log2PPM = NULL, signifdigits = 1)
}
\description{
Returns a rounded percentage given a log2 transformed PPM value
}
