% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_alpha_report.R
\name{make_alpha_report}
\alias{make_alpha_report}
\title{make_alpha_report(project=NULL, expvec=NULL, usefulexp=NULL, variable_list=NULL, measures=c("Observed", "InvSimpson"), cdict=NULL)}
\usage{
make_alpha_report(
  project = NULL,
  expvec = NULL,
  usefulexp = NULL,
  variable_list = NULL,
  measures = c("Observed", "InvSimpson"),
  cdict = NULL
)
}
\description{
Generates standard comparative plots for alpha diversity measures.
}
