% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_antibiogram_experiment.R
\name{make_antibiogram_experiment}
\alias{make_antibiogram_experiment}
\title{make_antibiogram_experiment(mgseqresfinder = NULL, SEobjresfinder = NULL)}
\usage{
make_antibiogram_experiment(mgseqresfinder = NULL, SEobjresfinder = NULL)
}
\description{
Agglomerates data from a resfinder SEobj or mgSeq object to return an antibiogram-style mgSeq or SummarizedExperiment object
}
