% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_classes_df.R
\name{make_classes_df}
\alias{make_classes_df}
\title{make_classes_df(curr_pt = NULL, compareby = NULL, wilcox_paired_by = NULL)}
\usage{
make_classes_df(curr_pt = NULL, compareby = NULL, wilcox_paired_by = NULL)
}
\description{
This function makes a data frame with sample to class mapping for use by calculate_matrix_stats.
}
