% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_featuredata_from_bedfile.R
\name{make_featuredata_from_bedfile}
\alias{make_featuredata_from_bedfile}
\title{make_featuredata_from_bedfile}
\usage{
make_featuredata_from_bedfile(opt = NULL, bedfile = NULL)
}
\description{
JAMSalpha function
}
