% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{make_featureplot}
\alias{make_featureplot}
\title{make_featureplot()}
\usage{
make_featureplot(
  dat = NULL,
  logtransbase = NULL,
  subsetby = NULL,
  compareby = NULL,
  colourby = NULL,
  shapeby = NULL,
  facetby = NULL,
  wrap_facet = FALSE,
  overlay_boxplot = FALSE
)
}
\description{
JAMSbeta wrapper for returning a ggplot boxplot, violin plot or scatterplot from a data frame. This function was concieved to work within other main plotting functions in JAMS. Do not feel frustrated if it does not work out of this context. My playground, my rules.
}
