% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap_report.R
\name{make_heatmap_report}
\alias{make_heatmap_report}
\title{make_heatmap_report(report=c("comparative", "exploratory", "PA"), project=NULL, expvec=NULL, usefulexp=NULL, featcutofftaxa=NULL, featcutofffunct=NULL, l2fchmtaxa=NULL, l2fchmfunc=NULL, applyfilters=c("stringent", "moderate", "none"), absolutemaxfeats=NULL, genomecompleteness=NULL, variable_list=NULL, adjustpval=TRUE, list.data=NULL, asPA=FALSE, mgSeqnorm=FALSE, cdict=NULL, showonlypbelow=NULL, makespreadsheets=FALSE, makeheatmaps=TRUE, maxnumheatmaps=NULL, ...)}
\usage{
make_heatmap_report(report = "comparative", project = NULL,
  expvec = NULL, usefulexp = NULL, appendtofilename = NULL,
  glomby = NULL, featcutofftaxa = NULL, featcutofffunct = NULL,
  l2fchmtaxa = NULL, l2fchmfunc = NULL, samplesToKeep = NULL,
  featuresToKeep = NULL, applyfilters = NULL,
  absolutemaxfeats = NULL, genomecompleteness = NULL,
  heatpalette = "smart", variable_list = NULL, adjustpval = NULL,
  list.data = NULL, hmasPA = FALSE, mgSeqnorm = FALSE,
  cdict = NULL, showonlypbelow = 0.05, makespreadsheets = FALSE,
  makeheatmaps = TRUE, maxnumheatmaps = NULL, numthreads = 4,
  nperm = 99, ...)
}
\description{
Generates standard comparative heatmaps and spreadsheet based on relative abundance difference between binary categories for analyses named in expvectors.
}
