% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_jams_file.R
\name{make_jams_file}
\alias{make_jams_file}
\title{make_jams_file}
\usage{
make_jams_file(
  opt = opt,
  outdir = opt$outdir,
  asRDS = TRUE,
  export_to_XL = FALSE
)
}
\description{
JAMSalpha function
}
