% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metadata_from_taxtable.R
\name{make_metadata_from_taxtable}
\alias{make_metadata_from_taxtable}
\title{make_metadata_from_taxtable(taxlvlinterest = NULL, taxainterest = NULL, nobs = TRUE,  metadatafn = NULL)}
\usage{
make_metadata_from_taxtable(taxlvlinterest = NULL, taxainterest = NULL,
  nobs = TRUE, metadatafn = NULL)
}
\description{
Generates JAMS-style metadata in xlsx format for comparing bacterial isolates which have been deposited in NCBI GenBank
}
