% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metagenomeSeq_experiments_bytaxon.R
\name{make_metagenomeSeq_experiments_bytaxon}
\alias{make_metagenomeSeq_experiments_bytaxon}
\title{make_metagenomeSeq_experiments_bytaxon(pheno=NULL, list.data=NULL, restricttotaxa=NULL)}
\usage{
make_metagenomeSeq_experiments_bytaxon(pheno = NULL, list.data = NULL,
  restricttotaxa = NULL, taxlevel = "LKT", onlyanalyses = NULL)
}
\description{
Returns a list vector of functional MetagenomeSeq MRexperiments for every analysis that is possible to make, restricted only to certain taxa.
}
