% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{make_phenolabels_from_phenotable}
\alias{make_phenolabels_from_phenotable}
\title{make_phenolabels_from_phenotable(phenotable = NULL)}
\usage{
make_phenolabels_from_phenotable(phenotable = NULL)
}
\description{
Does exactly what the name says.
}
