% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_readdata.R
\name{make_readdata}
\alias{make_readdata}
\title{make_readdata(Samples=NULL, list.data=NULL)}
\usage{
make_readdata(Samples = NULL, list.data = NULL)
}
\description{
This function makes a data frame with the sequencing statistics.
}
