% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_handling_functions.R
\name{make_sequences_caps}
\alias{make_sequences_caps}
\title{make_sequences_caps(sequence = NULL)}
\usage{
make_sequences_caps(sequence = NULL)
}
\description{
Transforms sequinr sequences to uppercase
}
