% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SummarizedExperiment_from_tables.R
\name{make_SummarizedExperiment_from_tables}
\alias{make_SummarizedExperiment_from_tables}
\title{make_SummarizedExperiment_from_tables(pheno = NULL, pheno_fn= NULL, counttable = NULL, counttable_fn = NULL, featuretable = NULL, featuretable_fn = NULL, onlysamples = NULL, analysisname = NULL, restricttoLKTs = NULL)}
\usage{
make_SummarizedExperiment_from_tables(
  pheno = NULL,
  pheno_fn = NULL,
  counttable = NULL,
  counttable_fn = NULL,
  featuretable = NULL,
  featuretable_fn = NULL,
  onlysamples = NULL,
  analysisname = NULL,
  restricttoLKTs = NULL
)
}
\description{
Safely makes a JAMS-compatible SummarizedExperiment object from a phenotable, a counttable and a featuretable.
}
