% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SummarizedExperiments.R
\name{make_SummarizedExperiments}
\alias{make_SummarizedExperiments}
\title{make_SummarizedExperiments(pheno = NULL, onlysamples = NULL,  onlyanalyses = NULL, minnumsampanalysis = NULL, minpropsampanalysis = 0.1, minPctFromCtg = NULL, minProbNumGenomes = NULL, restricttoLKTs = NULL, list.data = NULL)}
\usage{
make_SummarizedExperiments(
  pheno = NULL,
  onlysamples = NULL,
  onlyanalyses = NULL,
  minnumsampanalysis = NULL,
  minpropsampanalysis = 0.1,
  restricttoLKTs = NULL,
  list.data = NULL
)
}
\description{
Makes a SummarizedExperiment object for every analysis that is possible to make given loaded jams files in list.data.
}
