% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SummarizedExperiments_legacy.R
\name{make_SummarizedExperiments_legacy}
\alias{make_SummarizedExperiments_legacy}
\title{make_SummarizedExperiments_legacy(pheno = NULL, onlysamples = NULL,  onlyanalyses = c("LKT", "Product", "ECNumber", "Pfam", "Interpro", "resfinder", "PRINTS", "GO"), minnumsampanalysis = NULL, minpropsampanalysis = 0.1, restricttoLKTs = NULL, stratify_functions_by_taxon = TRUE, add_TNF_data = FALSE, list.data = NULL, phenolabels = NULL, cdict = cdict, threads = 8)}
\usage{
make_SummarizedExperiments_legacy(
  pheno = NULL,
  onlysamples = NULL,
  onlyanalyses = c("LKT", "Product", "ECNumber", "Pfam", "Interpro", "resfinder",
    "PRINTS", "GO"),
  minnumsampanalysis = NULL,
  minpropsampanalysis = 0.1,
  restricttoLKTs = NULL,
  stratify_functions_by_taxon = TRUE,
  add_TNF_data = FALSE,
  list.data = NULL,
  phenolabels = NULL,
  cdict = cdict,
  threads = 8
)
}
\description{
Makes a SummarizedExperiment object for every analysis that is possible to make given loaded jams files in list.data WITH JAMSFILES GENERATED USING VERSIONS OF JAMS < 2.0.0. FOR JAMSFILES GENERATED WITH DATABASES CREATED USING JAMS2 USE make_SummarizedExperiments.
}
