% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{maxRAMbytes_used}
\alias{maxRAMbytes_used}
\title{maxRAMbytes_used()
Wrapper for launching a report for a SINGLE analysis}
\usage{
maxRAMbytes_used()
}
\description{
maxRAMbytes_used()
Wrapper for launching a report for a SINGLE analysis
}
