% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_JAMS_SEobj.R
\name{merge_JAMS_SEobj}
\alias{merge_JAMS_SEobj}
\title{Safely merges two compatible JAMS-style SummarizedExperiment objects into a single JAMS-style SummarizedExperiment object. SE objects must belong to the same analysis type, and have no Sample names in common. Ideally they should have been generated with the same version of JAMS using the same JAMS database. If the latter condition is not met, the consequences are unforseen and you are ill-advised to proceed.}
\usage{
merge_JAMS_SEobj(ExpObj1 = NULL, ExpObj2 = NULL)
}
\arguments{
\item{ExpObj1}{a JAMS-style SummarizedExperiment object}

\item{ExpObj2}{a JAMS-style SummarizedExperiment object}
}
\description{
Safely merges two compatible JAMS-style SummarizedExperiment objects into a single JAMS-style SummarizedExperiment object. SE objects must belong to the same analysis type, and have no Sample names in common. Ideally they should have been generated with the same version of JAMS using the same JAMS database. If the latter condition is not met, the consequences are unforseen and you are ill-advised to proceed.
}
