% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sparse_matrix.R
\name{merge_sparse_matrix}
\alias{merge_sparse_matrix}
\title{Safely merges two sparse matrices. This function was written for exclusive use by the merge_JAMS_SEobj function (see). It has not been tested or benchmarked for a use other than that one.}
\usage{
merge_sparse_matrix(matlist = NULL)
}
\arguments{
\item{matlist}{- A list of sparse matrices}
}
\description{
Safely merges two sparse matrices. This function was written for exclusive use by the merge_JAMS_SEobj function (see). It has not been tested or benchmarked for a use other than that one.
}
