% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa2JAMS.R
\name{mpa2JAMS}
\alias{mpa2JAMS}
\title{mpa2JAMS(mpa_folder = NULL, only_prefixes = NULL, output_folder = NULL, mpa_file_suffix = "_profile.txt", return_mpa_list.data = TRUE, verbose = TRUE)}
\usage{
mpa2JAMS(
  mpa_folder = NULL,
  only_prefixes = NULL,
  output_folder = NULL,
  mpa_file_suffix = "_profile.txt",
  return_mpa_list.data = TRUE,
  verbose = TRUE
)
}
\description{
Makes .jams files from Metaphlann output files. This function is experimental, use at your own risk and peril.
}
