% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_subsetting_sample_selector.R
\name{multiple_subsetting_sample_selector}
\alias{multiple_subsetting_sample_selector}
\title{multiple_subsetting_sample_selector(SEobj = NULL, phenotable = NULL, subsetby = NULL, compareby = NULL, cats_to_ignore = "N_A")}
\usage{
multiple_subsetting_sample_selector(
  SEobj = NULL,
  phenotable = NULL,
  subsetby = NULL,
  compareby = NULL,
  cats_to_ignore = "N_A"
)
}
\description{
Application for subsetting a phenotable into up to three tiers of subsets within subsets. Returns a named list. The first element, named "Subsets_stats" is a data frame of the compareby stats within each specific subset. Subsequent list named items are vectors of the sample names within each subset specified by the "Subset_Tier_Class_Name" column of the "Subsets_stats" data frame.
}
