% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_samples.R
\name{name_samples}
\alias{name_samples}
\title{name_samples(list.data=NULL)}
\usage{
name_samples(list.data = NULL)
}
\description{
Given a list.data object, returns a vector of sample names present in the object.
}
