% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_dada2_amplicon_sizes.R
\name{obtain_dada2_amplicon_sizes}
\alias{obtain_dada2_amplicon_sizes}
\title{obtain_dada2_amplicon_sizes(mergers_list = NULL, sample_name = NULL, fun_to_apply = "median")}
\usage{
obtain_dada2_amplicon_sizes(
  mergers_list = NULL,
  sample_name = NULL,
  fun_to_apply = "median"
)
}
\description{
This is a JAMS16 function.
This function will return the statistical transformation set by fun_to_apply on a given sample's merged F-R pairs within a DADA2 "mergers" object.
}
