% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_abundance.R
\name{pareto_abundance}
\alias{pareto_abundance}
\title{pareto_abundance(LKTdose=NULL, samplename=NULL, taxlevel=NULL, assemblystats=NULL, showcompletenesstext=TRUE, showcompletenessline=TRUE, showcontigline=FALSE, showparetocurve=TRUE, showpercentagetext=TRUE)}
\usage{
pareto_abundance(
  LKTdose = NULL,
  samplename = NULL,
  list.data = NULL,
  taxlevel = NULL,
  assemblystats = NULL,
  showcompletenesstext = TRUE,
  showcompletenessline = TRUE,
  showcontigline = FALSE,
  showparetocurve = TRUE,
  showpercentagetext = TRUE
)
}
\description{
Plots a pareto chart of the relative abundance of a taxon.
}
