% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_graph.R
\name{plot_bar_graph}
\alias{plot_bar_graph}
\title{Creates bar graph of relative taxonomic abundance from SummarizedExperiment object.}
\usage{
plot_bar_graph(
  ExpObj = NULL,
  glomby = NULL,
  absolute = FALSE,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  groupby = NULL,
  label_samples = FALSE,
  colourby = NULL,
  subsetby = NULL,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  PPM_normalize_to_bases_sequenced = FALSE,
  total_units = 100,
  threshold = 2,
  cat_pos = -2,
  cat_text_size = 3,
  legend_text_size = 4,
  border_color = "white",
  cdict = NULL,
  feature_cdict = NULL,
  addtit = NULL,
  grid = TRUE,
  ignoreunclassified = FALSE,
  class_to_ignore = "N_A",
  ...
)
}
\description{
Creates bar graph of relative taxonomic abundance from SummarizedExperiment object.
}
