% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bias.R
\name{plot_bias}
\alias{plot_bias}
\title{plot_bias(bias, method = c("pca", tSNE"), frac = 0.25))}
\usage{
plot_bias(bias = NULL, method = "pca", frac = 0.25)
}
\description{
Makes biased plots. Just kidding.
}
