% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_duet_heatmap.R
\name{plot_duet_heatmap}
\alias{plot_duet_heatmap}
\title{plot_duet_heatmap(mgseqobj = NULL, glomby = NULL, colcategories = NULL, compareby = NULL, adjustpval = NULL, showl2fc = TRUE, showpval = TRUE, stattype = "spearman", subsetby = NULL, maxnumfeatallowed = 10000, minabscorrcoeff = NULL, ntopvar = NULL, featmaxatleastPPM = 0, featcutoff = c(0, 0), applyfilters = NULL, featuresToKeep = NULL, samplesToKeep = NULL, genomecompleteness = NULL, list.data = NULL, showGram = TRUE, showphylum = TRUE, addtit = NULL, mgSeqnorm = FALSE, cdict = NULL, ignoreunclassified = TRUE, class_to_ignore = NULL)}
\usage{
plot_duet_heatmap(
  mgseqobj = NULL,
  glomby = NULL,
  colcategories = NULL,
  compareby = NULL,
  adjustpval = NULL,
  showl2fc = TRUE,
  showpval = TRUE,
  stattype = "spearman",
  subsetby = NULL,
  maxnumfeatallowed = 10000,
  minabscorrcoeff = NULL,
  ntopvar = NULL,
  featmaxatleastPPM = 0,
  featcutoff = c(0, 0),
  applyfilters = NULL,
  featuresToKeep = NULL,
  samplesToKeep = NULL,
  genomecompleteness = NULL,
  list.data = NULL,
  showGram = TRUE,
  showphylum = TRUE,
  addtit = NULL,
  mgSeqnorm = FALSE,
  cdict = NULL,
  ignoreunclassified = TRUE,
  class_to_ignore = NULL
)
}
\description{
Plots correlation heatmaps annotated by the metadata or a correlelogram of features
DISCLAIMER: This function is experimental
}
