% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_relabund.R
\name{plot_feature_relabund}
\alias{plot_feature_relabund}
\title{plot_feature_relabund<-function(ExpObj=NULL, mgSeqnorm=FALSE, feature=NULL, glomby=NULL, accession=NULL, groupby=NULL, colourby=NULL, shapeby=NULL, subsetby=NULL, uselog=TRUE, statmeth="wilcox.test", samplesToKeep=NULL, featuresToKeep=NULL, signiflabel=c("p.signif", "p.format"), plottitle=NULL, asPA=FALSE, subsetbytaxlevel=NULL, taxtable=NULL, list.data=NULL, cdict=NULL, ...)}
\usage{
plot_feature_relabund(
  ExpObj = NULL,
  mgSeqnorm = FALSE,
  feature = NULL,
  glomby = NULL,
  accession = NULL,
  groupby = NULL,
  colourby = NULL,
  shapeby = NULL,
  subsetby = NULL,
  uselog = TRUE,
  statmeth = "wilcox.test",
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  signiflabel = "p.format",
  plottitle = NULL,
  asPA = FALSE,
  subsetbytaxlevel = NULL,
  taxtable = NULL,
  list.data = NULL,
  cdict = NULL,
  max_categories = 3,
  ...
)
}
\description{
Plots the relative abundance of an entity present in a sample or group of samples and subsets by taxonomy.
}
