% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_relabund_devel.R
\name{plot_feature_relabund_devel}
\alias{plot_feature_relabund_devel}
\title{plot_feature_relabund_devel()}
\usage{
plot_feature_relabund_devel(mgseqobj = NULL, glomby = NULL,
  heatpalette = "smart", stattype = "spearman", hmasPA = FALSE,
  compareby = NULL, ntop = NULL, cluster_rows = FALSE,
  subsetby = NULL, applyfilters = NULL, featmaxatleastPPM = 0,
  featcutoff = c(0, 0), samplesToKeep = NULL, featuresToKeep = NULL,
  adjustpval = FALSE, showonlypbelow = NULL, showpval = TRUE,
  showcorrcoeff = TRUE, minabscorrcoeff = NULL,
  genomecompleteness = NULL, list.data = NULL, addtit = NULL,
  mgSeqnorm = FALSE, cdict = NULL, numthreads = 4, nperm = 99,
  statsonlog = TRUE, ignoreunclassified = TRUE, returnstats = FALSE,
  ...)
}
\description{
Plots correlation heatmaps annotated by the metadata or a correlelogram of features
}
