% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_taxonomy_donut.R
\name{plot_genome_taxonomy_donut}
\alias{plot_genome_taxonomy_donut}
\title{plot_genome_taxonomy_donut(opt = opt, contigsdata = NULL)}
\usage{
plot_genome_taxonomy_donut(
  opt = opt,
  assemblystats = NULL,
  taxlevel = "LKT",
  isolatename = NULL,
  percentage_threshold = 99,
  ntop = 8,
  brewer_palette = 3
)
}
\description{
JAMSalpha function
}
