% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_readscatter.R
\name{plot_readscatter}
\alias{plot_readscatter}
\title{plot_readscatter(readdata=NULL, phenotable=NULL, discrete=NULL)}
\usage{
plot_readscatter(readdata = NULL, phenotable = NULL, discrete = NULL)
}
\description{
Makes a scatterplot of percentage of reads assembled into contigs by sequencing depth.
}
