% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relabund_heatmap_mgseq.R
\name{plot_relabund_heatmap_mgseq}
\alias{plot_relabund_heatmap_mgseq}
\title{plot_relabund_heatmap_mgseq(mgseqobj = NULL, glomby = NULL, heatpalette = "smart", hmtype = NULL, hmasPA = FALSE, compareby = NULL, invertbinaryorder = FALSE, ntop = NULL, ordercolsby = NULL, cluster_samples_per_heatmap = FALSE, colcategories = NULL, cluster_rows = TRUE, subsetby = NULL, applyfilters = NULL, featmaxatleastPPM = 0, featcutoff = c(0, 0), samplesToKeep = NULL, featuresToKeep = NULL, adjustpval = FALSE, showonlypbelow = NULL, showpval = TRUE, showl2fc = TRUE, maxl2fc = NULL, minl2fc = NULL, genomecompleteness = NULL, list.data = NULL, addtit = NULL,  scaled = FALSE, mgSeqnorm = FALSE, cdict = NULL, maxnumheatmaps = NULL, numthreads = 1, nperm = 99, statsonlog = TRUE, ignoreunclassified = TRUE, returnstats = FALSE, class_to_ignore = NULL, ...)}
\usage{
plot_relabund_heatmap_mgseq(
  mgseqobj = NULL,
  glomby = NULL,
  heatpalette = "smart",
  hmtype = NULL,
  hmasPA = FALSE,
  compareby = NULL,
  invertbinaryorder = FALSE,
  ntop = NULL,
  ordercolsby = NULL,
  cluster_samples_per_heatmap = FALSE,
  colcategories = NULL,
  cluster_rows = TRUE,
  subsetby = NULL,
  applyfilters = NULL,
  featmaxatleastPPM = 0,
  featcutoff = c(0, 0),
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  adjustpval = FALSE,
  showonlypbelow = NULL,
  showpval = TRUE,
  showl2fc = TRUE,
  maxl2fc = NULL,
  minl2fc = NULL,
  genomecompleteness = NULL,
  list.data = NULL,
  addtit = NULL,
  scaled = FALSE,
  mgSeqnorm = FALSE,
  cdict = NULL,
  maxnumheatmaps = NULL,
  numthreads = 1,
  nperm = 99,
  statsonlog = TRUE,
  ignoreunclassified = TRUE,
  returnstats = FALSE,
  class_to_ignore = NULL,
  ...
)
}
\description{
Plots relative abundance heatmaps annotated by the metadata
}
