% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{plot_table_a4_autofit_paged}
\alias{plot_table_a4_autofit_paged}
\title{plot_table_a4_autofit_paged(dataframe = NULL, title = NULL, orientation = NULL, rows_per_page = 35, start_fontsize = 10, min_fontsize = 5, margins_portrait_mm = c(top=6, right=6, bottom=6, left=6), margins_landscape_mm = c(top=6, right=8, bottom=6, left=8), gap_mm = 2, cell_padding_mm = c(v = 0.6, h = 1.2), header_padding_mm = c(v = 0.8, h = 1.2), lineheight = 0.95, wrap = FALSE, wrap_chars = NULL, wrap_min = 12, wrap_max = 60)
EXPERIMENTAL novel plotting function for printing data frames into a pdf.
Usage:
pdf("my_landscape.pdf", height = 210/25.4, width = 297/25.4)  # A4 landscape
plot_table_a4_autofit_paged(dataframe = opt$featuredata[1:80, ], title="Featuretable", rows_per_page = 35, wrap = TRUE, wrap_chars = 20, start_fontsize = 15, min_fontsize = 2,  orientation = "landscape")
dev.off()}
\usage{
plot_table_a4_autofit_paged(
  dataframe = NULL,
  title = NULL,
  orientation = "landscape",
  rows_per_page = 35,
  start_fontsize = 10,
  min_fontsize = 5,
  margins_portrait_mm = c(top = 6, right = 6, bottom = 6, left = 6),
  margins_landscape_mm = c(top = 6, right = 8, bottom = 6, left = 8),
  gap_mm = 2,
  cell_padding_mm = c(v = 0.6, h = 1.2),
  header_padding_mm = c(v = 0.8, h = 1.2),
  lineheight = 0.95,
  wrap = FALSE,
  wrap_chars = NULL,
  wrap_min = 12,
  wrap_max = 60
)
}
\description{
plot_table_a4_autofit_paged(dataframe = NULL, title = NULL, orientation = NULL, rows_per_page = 35, start_fontsize = 10, min_fontsize = 5, margins_portrait_mm = c(top=6, right=6, bottom=6, left=6), margins_landscape_mm = c(top=6, right=8, bottom=6, left=8), gap_mm = 2, cell_padding_mm = c(v = 0.6, h = 1.2), header_padding_mm = c(v = 0.8, h = 1.2), lineheight = 0.95, wrap = FALSE, wrap_chars = NULL, wrap_min = 12, wrap_max = 60)
EXPERIMENTAL novel plotting function for printing data frames into a pdf.
Usage:
pdf("my_landscape.pdf", height = 210/25.4, width = 297/25.4)  # A4 landscape
plot_table_a4_autofit_paged(dataframe = opt$featuredata[1:80, ], title="Featuretable", rows_per_page = 35, wrap = TRUE, wrap_chars = 20, start_fontsize = 15, min_fontsize = 2,  orientation = "landscape")
dev.off()
}
