% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxsplit_tree.R
\name{plot_taxsplit_tree}
\alias{plot_taxsplit_tree}
\title{plot_taxsplit_tree()}
\usage{
plot_taxsplit_tree(
  taxsplit_df = NULL,
  strat_tt = NULL,
  layout = c("rectangular", "fan", "circular"),
  open.angle = 5,
  tip_offset = NULL,
  fruit_offset = NULL,
  fruit_pwidth = NULL,
  bar_width = 0.85,
  tip_size = NULL,
  show_prevalence = TRUE,
  bar_fill = "phylum",
  low_col = "#fde0dd",
  high_col = "#7a0177",
  hilight_rank = "Phylum",
  taxon_hilight_palette = NULL,
  hilight_alpha = 0.4,
  tree_size = 0.4,
  legend_position = "right",
  plotitstrat = NULL
)
}
\description{
Helper plotting function for use exclusively within plot_relabund_features (q.v.). Do not attempt to use this outside of this context, or you will find yourself in a frightful muddle.
}
