% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Venn.R
\name{plot_Venn}
\alias{plot_Venn}
\title{Creates Venn Diagram (2, 3, or 4 way) from metagenomeSeq object and returns list of features broken down from diagram}
\usage{
plot_Venn(
  mgseqobj = NULL,
  category = NULL,
  glomby = NULL,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  mgSeqnorm = FALSE,
  featmaxatleastPPM = 0,
  featcutoff = c(0, 0),
  aggFun = rowMeans,
  title = "",
  colours = NULL,
  labels = NULL,
  pdf = NULL,
  ...
)
}
\arguments{
\item{mgseqobj}{a metagenomeSeq object.}

\item{category}{string giving pData field to group samples by.}

\item{glomby}{string giving taxonomic level (from tax_table) to plot at.}

\item{samplesToKeep}{vector with samples to plot.}

\item{featuresToKeep}{vector with features to plot.}

\item{mgSeqnorm}{whether to normalize counts.}

\item{featmaxatleastPPM}{only count features with at least a certain PPM at max value.}

\item{featcutoff}{only count features with at least a certain ppm in a certain percentage of samples.}

\item{title}{plot title.}

\item{colours}{use list of colors.}

\item{labels}{to overwrite metagenomeseq names}

\item{pdf}{write plot as a pdf to a certain file.}
}
\description{
Creates Venn Diagram (2, 3, or 4 way) from metagenomeSeq object and returns list of features broken down from diagram
}
\examples{
function plot_Venn(mgseqobj = NULL,category = NULL, glomby = NULL, samplesToKeep = NULL, featuresToKeep = NULL, mgSeqnorm = FALSE, featmaxatleastPPM = 0, featcutoff = c(0, 0), aggFun = rowMeans, title = "", colours= NULL, pdf = NULL, ...)
}
