% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{post_message}
\alias{post_message}
\title{post_message(msg = NULL, face = "plain", just = "centre", color = "black", size = 20)}
\usage{
post_message(
  msg = NULL,
  face = "plain",
  just = "centre",
  color = "black",
  size = 20
)
}
\description{
Wrapper function for plotting text to a full page on a pdf.
}
