% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_contigs_for_JAMS.R
\name{prepare_contigs_for_JAMS}
\alias{prepare_contigs_for_JAMS}
\title{prepare_contigs_for_JAMS}
\usage{
prepare_contigs_for_JAMS(opt = NULL, fastafile = NULL, contig_minlength = 500)
}
\description{
JAMSalpha function
}
