% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_jams_report.R
\name{print_jams_report}
\alias{print_jams_report}
\title{print_jams_report(opt=opt)}
\usage{
print_jams_report(
  opt = opt,
  outputdir = NULL,
  elements = c("readplots", "taxonomy", "SixteenSid", "resfinder", "abricate",
    "plasmidfinder", "vfdb", "func"),
  plotucobias = FALSE
)
}
\description{
JAMSalpha function
}
