% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{print_table(tb=NULL, tabletitle=NULL, fontsize=6, numrows=20)}
\usage{
print_table(tb = NULL, tabletitle = NULL, fontsize = 6,
  numrows = 20)
}
\description{
JAMSalpha function
}
