% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{quiet}
\alias{quiet}
\title{quiet(function(x))
Suppresses output messages
By Hadley Wickham
http://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html}
\usage{
quiet(x)
}
\description{
quiet(function(x))
Suppresses output messages
By Hadley Wickham
http://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html
}
