% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{RAMbytes_status}
\alias{RAMbytes_status}
\title{RAMbytes_status(RAMbytesavail = NULL)
Reports how much RAM memory is being or has maximally been used.}
\usage{
RAMbytes_status(RAMbytesavail = NULL)
}
\description{
RAMbytes_status(RAMbytesavail = NULL)
Reports how much RAM memory is being or has maximally been used.
}
