% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_experiment.R
\name{randomize_experiment}
\alias{randomize_experiment}
\title{randomize_experiment(mgseqobj = NULL, newname = "Noise", shuffleclassifieds = FALSE, shufflemetadata = TRUE, iterations = 10)}
\usage{
randomize_experiment(mgseqobj = NULL, newname = "Noise",
  shuffleclassifieds = FALSE, shufflemetadata = TRUE,
  iterations = 10)
}
\description{
Shuffles data (counts and metadata) of a metagenomeSeq MRexperiment object. Useful for sanity checking.
}
