% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_completeness_to_CheckM_style.R
\name{reformat_completeness_to_CheckM_style}
\alias{reformat_completeness_to_CheckM_style}
\title{reformat_completeness_to_CheckM_style(completeness_df = NULL)}
\usage{
reformat_completeness_to_CheckM_style(completeness_df = NULL)
}
\description{
JAMSalpha function
}
