% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_sequences_consecutively.R
\name{rename_sequences_consecutively}
\alias{rename_sequences_consecutively}
\title{rename_sequences_consecutively(sequence=NULL, sequence=NULL, headerprefix="ctg")}
\usage{
rename_sequences_consecutively(sequence = NULL, headerprefix = "ctg")
}
\description{
Renames sequences numerically consecutively
}
