% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{replace_NAs_with_character}
\alias{replace_NAs_with_character}
\title{replace_NAs_with_character(df)
Replaces NAs with a specific character in a data frame}
\usage{
replace_NAs_with_character(df = NULL, replacement = "N_A")
}
\description{
replace_NAs_with_character(df)
Replaces NAs with a specific character in a data frame
}
