% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_Sample_names_in_SEobj.R
\name{replace_Sample_names_in_SEobj}
\alias{replace_Sample_names_in_SEobj}
\title{replace_Sample_names_in_SEobj(curr_SEobj = NULL, Sample_name_mapping = NULL, oldname_column = "Sample", newname_column = NULL)}
\usage{
replace_Sample_names_in_SEobj(
  curr_SEobj = NULL,
  Sample_name_mapping = NULL,
  oldname_column = "Sample",
  newname_column = NULL
)
}
\description{
Renames samples in JAMS SummarizedExperiment objects safely.
Use with care and at your own risk
}
