% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{retrieve_GCA_accession_URL}
\alias{retrieve_GCA_accession_URL}
\title{retrieve_GCA_accession_URL(assemblytargetname = NULL)}
\usage{
retrieve_GCA_accession_URL(assemblytargetname)
}
\description{
Given a single valid GCA assembly accession number, this function will return the https URL for downloading the fasta.gz file for that assembly accession.
}
