% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{shrink_perbasecoverage}
\alias{shrink_perbasecoverage}
\title{shrink_perbasecoverage(perbasecoverage=NULL, percentage=2)}
\usage{
shrink_perbasecoverage(perbasecoverage = NULL, percentage = 2)
}
\description{
#Subset down to n% of the dataset to make it calculatable. To be used only by JAMSalpha.
}
