% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_contigs_into_bins.R
\name{sort_contigs_into_bins}
\alias{sort_contigs_into_bins}
\title{sort_contigs_into_bins(opt = NULL)}
\usage{
sort_contigs_into_bins(opt = NULL, sortedbamfile = NULL)
}
\description{
JAMSalpha function
}
