% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{spew_heatmap_report}
\alias{spew_heatmap_report}
\title{spew_heatmap_report(c("analysis", "report"))
Wrapper for launching a report for a SINGLE analysis}
\usage{
spew_heatmap_report(hmcomb)
}
\description{
spew_heatmap_report(c("analysis", "report"))
Wrapper for launching a report for a SINGLE analysis
}
