% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{split_featname}
\alias{split_featname}
\title{split_featname(featname = NULL, thresh_featname_split = 40)}
\usage{
split_featname(featname = NULL, thresh_featname_split = 40)
}
\description{
For a feature name (or any string, for that matter) will add a carriage return to the middle of the string if it exceeds the threshold number of characters. This is used in the heatmap function.
}
