% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SraRunTable_to_metadata.R
\name{SraRunTable_to_metadata}
\alias{SraRunTable_to_metadata}
\title{SraRunTable_to_metadata(SraRunTable=NULL, Sample=NULL, discrete=NULL, subsettable=NULL, continuous=NULL)}
\usage{
SraRunTable_to_metadata(SraRunTable = NULL, Sample = "Run",
  discrete = NULL, subsettable = NULL, continuous = NULL)
}
\description{
Converts an SRA Run Table to JAMS-style metadata (phenotable, phenolabels).
}
