% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{tally_metadata}
\alias{tally_metadata}
\title{tally_metadata(md = NULL, column_to_group_by = NULL, columns_to_tally = NULL)}
\usage{
tally_metadata(md = NULL, column_to_group_by = NULL, columns_to_tally = NULL)
}
\description{
Tallies the number of classes present in a metadata df and returns a data frame with counts.
}
