% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{update_ExpObj_metadata}
\alias{update_ExpObj_metadata}
\title{update_ExpObj_metadata(ExpObj = NULL, phenotable = NULL)}
\usage{
update_ExpObj_metadata(ExpObj = NULL, phenotable = NULL)
}
\description{
Updates phenotable of a SummarizedExperiment object
}
