% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{update_metadata}
\usage{
update_metadata(metadataXL = NULL, phenotable = NULL,
  phenolabels = NULL, expvec = NULL, maxclass = 10, minGbNAHS = 0,
  minPctAssembly = 0, readdata = NULL)
}
\description{
Updates the metadata in from an MetagenomeSeq experiments present in expvec from an Excel spreadsheet with phenotable and phenolabels.
}
