% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{whoopsieplot}
\alias{whoopsieplot}
\title{whoopsieplot(msg = NULL)
Shuts down the device and gives a message on error, useful for when making reports.}
\usage{
whoopsieplot(msg = "trying to do this.")
}
\description{
whoopsieplot(msg = NULL)
Shuts down the device and gives a message on error, useful for when making reports.
}
