% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_handling_functions.R
\name{write_contigs_to_system}
\alias{write_contigs_to_system}
\title{write_contigs_to_system(opt = NULL, contig_names = NULL, filename = NULL)
Wrapper for writing a list of contigs from opt$contigsdata to a single multifasta file.}
\usage{
write_contigs_to_system(opt = NULL, contig_names = NULL, filename = NULL)
}
\description{
JAMSalpha function
}
