% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_phenotable.R
\name{adjust_phenotable}
\alias{adjust_phenotable}
\title{adjust_phenotable(phenotable=NULL, phenolabels=NULL, readdata=NULL, list.data=list.data)}
\usage{
adjust_phenotable(phenotable = NULL, phenolabels = NULL,
  readdata = NULL, list.data = NULL, addtaxlevelstoisolates = NULL)
}
\description{
Adjusts the metadata table into a format which can be used by the MetagenomeSeq package.
}
