% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_matrix_stats.R
\name{calculate_matrix_stats}
\alias{calculate_matrix_stats}
\title{calculate_matrix_stats(countmatrix = NULL, uselog = NULL, statsonlog = TRUE, stattype = NULL, classesvector = NULL, invertbinaryorder = FALSE, fun_for_l2fc = "geom_mean", threshPA = 0, numthreads = 1, nperm = 999)}
\usage{
calculate_matrix_stats(
  countmatrix = NULL,
  uselog = NULL,
  statsonlog = TRUE,
  stattype = NULL,
  classesvector = NULL,
  invertbinaryorder = FALSE,
  fun_for_l2fc = "geom_mean",
  threshPA = 0,
  numthreads = 1,
  nperm = 999
)
}
\description{
Returns a data frame with the statistics for a feature count matrix ordered by highest variance or lowest Mann-Whitney-Wilcoxon test between binary categories.
}
