% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_handling_functions.R
\name{countfastq}
\alias{countfastq}
\title{countfastq(fastqfile)
Wrapper for counting number of reads and number of bases of a fastq file in the system.
Returns a vector with read counts and base counts.
Caveat: Input fastq file must have exactly four lines per sequence. If your fastq files do not fit this criterion, you are totally bonkers.}
\usage{
countfastq(fastqfile)
}
\description{
countfastq(fastqfile)
Wrapper for counting number of reads and number of bases of a fastq file in the system.
Returns a vector with read counts and base counts.
Caveat: Input fastq file must have exactly four lines per sequence. If your fastq files do not fit this criterion, you are totally bonkers.
}
