% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_handling_functions.R
\name{countfastq_files}
\alias{countfastq_files}
\title{countfastq_files(fastqfiles = NULL, threads = NULL)
Wrapper for applying the countfastq() function to a vector of filenames using multiple threads.
Returns a dataframe with read counts and base counts for each input fastq file.}
\usage{
countfastq_files(fastqfiles = NULL, threads = NULL)
}
\description{
countfastq_files(fastqfiles = NULL, threads = NULL)
Wrapper for applying the countfastq() function to a vector of filenames using multiple threads.
Returns a dataframe with read counts and base counts for each input fastq file.
}
