% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{declare_filtering_presets}
\alias{declare_filtering_presets}
\title{declare_filtering_presets(analysis = NULL, applyfilters = NULL, featcutoff = NULL, GenomeCompletenessCutoff = NULL, PctFromCtgscutoff = NULL, maxl2fc = NULL, minl2fc = NULL)}
\usage{
declare_filtering_presets(
  analysis = NULL,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL,
  maxl2fc = NULL,
  minl2fc = NULL
)
}
\description{
Performs vetting of a SummarizedExperiment object for use in several functions
}
