% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{estimate_genome_completeness}
\alias{estimate_genome_completeness}
\title{estimate_genome_completeness(contigs2length)
JAMSalpha function to estimate genome completeness from contigs2length}
\usage{
estimate_genome_completeness(contigs2length = NULL)
}
\description{
estimate_genome_completeness(contigs2length)
JAMSalpha function to estimate genome completeness from contigs2length
}
