% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpObjVetting.R
\name{ExpObjVetting}
\alias{ExpObjVetting}
\title{ExpObjVetting(ExpObj = NULL)}
\usage{
ExpObjVetting(
  ExpObj = NULL,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  glomby = NULL,
  variables_to_fix = NULL,
  class_to_ignore = NULL
)
}
\description{
Performs vetting of a SummarizedExperiment object for use in several functions
}
