% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_expvec_to_XL.R
\name{export_expvec_to_XL}
\alias{export_expvec_to_XL}
\title{export_expvec_to_XL(expvec = NULL, usefulexp = NULL, filename = NULL, asPPM = TRUE, PPM_normalize_to_bases_sequenced = FALSE, applyfilters = NULL, featcutoff = NULL, GenomeCompletenessCutoff = NULL, PctFromCtgscutoff = NULL, includemetadata = TRUE, returncounts = FALSE)}
\usage{
export_expvec_to_XL(
  expvec = NULL,
  usefulexp = NULL,
  filename = NULL,
  asPPM = TRUE,
  PPM_normalize_to_bases_sequenced = FALSE,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL,
  includemetadata = TRUE,
  returncounts = FALSE
)
}
\description{
Exports counts and featuredata in a SummarizedExperiment vector into a single spreadsheet.
}
