% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_experiment.R
\name{filter_experiment}
\alias{filter_experiment}
\title{filter_experiment(ExpObj = NULL, featmaxatleastPPM = 0, featcutoff = c(0, 0), samplesToKeep = NULL, featuresToKeep = NULL, asPPM = TRUE, PPM_normalize_to_bases_sequenced = FALSE, GenomeCompletenessCutoff = NULL, PctFromCtgscutoff = NULL, mgSeqnorm = FALSE)}
\usage{
filter_experiment(
  ExpObj = NULL,
  featmaxatleastPPM = 0,
  featcutoff = c(0, 0),
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  asPPM = TRUE,
  PPM_normalize_to_bases_sequenced = FALSE,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL,
  mgSeqnorm = FALSE
)
}
\description{
Filters a SummarizedExperiment or metagenomeSeq object by several criteria.
}
