% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_experiment.R
\name{filter_experiment}
\alias{filter_experiment}
\title{filter_experiment(mgseqobj=NULL, featmaxatleastPPM=0, featcutoff=c(0, 0), samplesToKeep=NULL, featuresToKeep=NULL, asPA=FALSE, asPPM=TRUE, mgSeqnorm=FALSE)}
\usage{
filter_experiment(mgseqobj = NULL, featmaxatleastPPM = 0,
  featcutoff = c(0, 0), samplesToKeep = NULL, featuresToKeep = NULL,
  asPA = FALSE, asPPM = TRUE, mgSeqnorm = FALSE)
}
\description{
Filters a metagenomeSeq object by several criteria.
}
