% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_experiment.R
\name{filter_experiment}
\alias{filter_experiment}
\title{filter_experiment(ExpObj = NULL, featmaxatleastPPM = 0, featcutoff = c(0, 0), discard_SDoverMean_below = NULL, samplesToKeep = NULL, featuresToKeep = NULL, normalization = "relabund", asPPM = TRUE, PPM_normalize_to_bases_sequenced = FALSE, flush_out_empty_samples = FALSE, GenomeCompletenessCutoff = NULL, PctFromCtgscutoff = NULL)}
\usage{
filter_experiment(
  ExpObj = NULL,
  featmaxatleastPPM = 0,
  featcutoff = c(0, 0),
  discard_SDoverMean_below = NULL,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  normalization = "relabund",
  asPPM = TRUE,
  PPM_normalize_to_bases_sequenced = FALSE,
  flush_out_empty_samples = FALSE,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL
)
}
\description{
Filters a SummarizedExperiment object by several criteria.
}
