% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contig_coverage.R
\name{get_contig_coverage}
\alias{get_contig_coverage}
\title{get_contig_coverage}
\usage{
get_contig_coverage(opt = NULL, markduplicates = FALSE,
  classifyunassembled = TRUE)
}
\description{
JAMSalpha function
}
