% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting_functions.R
\name{harvest_functions}
\alias{harvest_functions}
\title{harvest_functions}
\usage{
harvest_functions(
  opt = opt,
  noninterproanalyses = c("FeatType", "ECNumber", "Product", "resfinder",
    "plasmidfinder", "napdos", "serofinderH", "serofinderO", "vfdb", "abricate")
)
}
\description{
JAMSalpha function
}
