% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_alpha_report.R
\name{make_alpha_report}
\alias{make_alpha_report}
\title{make_alpha_report(project=NULL, expvec=NULL, usefulexp=NULL, variable_list=NULL, measures=c("Observed", "InvSimpson"), cdict=NULL)}
\usage{
make_alpha_report(
  project = NULL,
  expvec = NULL,
  usefulexp = NULL,
  appendtofilename = NULL,
  variable_list = NULL,
  measures = c("Observed", "InvSimpson", "GeneCounts"),
  cdict = NULL,
  stratify_by_kingdoms = TRUE,
  glomby = NULL,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL,
  PPM_normalize_to_bases_sequenced = FALSE,
  addtit = NULL,
  max_pairwise_cats = 4,
  ignoreunclassified = TRUE,
  class_to_ignore = "N_A",
  makespreadsheets = FALSE,
  makeplots = TRUE,
  ...
)
}
\description{
Generates standard comparative plots for alpha diversity measures.
}
