% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_colour_dictionary.R
\name{make_colour_dictionary}
\alias{make_colour_dictionary}
\title{make_colour_dictionary(variable_list=NULL, pheno=NULL)}
\usage{
make_colour_dictionary(variable_list = NULL, pheno = NULL)
}
\description{
Returns a dictionary of colours for each class within each variable_list
}
