% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_colour_dictionary.R
\name{make_colour_dictionary}
\alias{make_colour_dictionary}
\title{make_colour_dictionary(variable_list = NULL, pheno = NULL, phenolabels = NULL, class_to_ignore = "N_A", colour_of_class_to_ignore = "#bcc2c2", colour_table = NULL, shuffle = FALSE, print_palette = FALSE)}
\usage{
make_colour_dictionary(
  variable_list = NULL,
  pheno = NULL,
  phenolabels = NULL,
  class_to_ignore = "N_A",
  colour_of_class_to_ignore = "#bcc2c2",
  colour_table = NULL,
  shuffle = FALSE,
  print_palette = FALSE
)
}
\description{
Returns a dictionary of colours for each class within each variable_list or phenotable
}
