% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_jams_file.R
\name{make_jams_file}
\alias{make_jams_file}
\title{make_jams_file}
\usage{
make_jams_file(
  opt = opt,
  objectstoinclude = c("contigsdata", "featuredata", "abundances",
    "taxonomic_Quality_split_list", opt$blastanalyses, "bam_contig_depths",
    "TNF_contigs", "readstats", "fastqstats", "projinfo"),
  outdir = opt$outdir
)
}
\description{
JAMSalpha function
}
