% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metagenomeSeq_experiments.R
\name{make_metagenomeSeq_experiments}
\alias{make_metagenomeSeq_experiments}
\title{make_metagenomeSeq_experiments(pheno = NULL, onlysamples = NULL,  onlyanalyses = NULL, minnumsampanalysis = NULL, minpropsampanalysis = 0.1, minPctFromCtg = NULL, minProbNumGenomes = NULL, restricttoLKTs = NULL, list.data = NULL)}
\usage{
make_metagenomeSeq_experiments(pheno = NULL, onlysamples = NULL,
  onlyanalyses = NULL, minnumsampanalysis = NULL,
  minpropsampanalysis = 0.1, minPctFromCtg = NULL,
  minProbNumGenomes = NULL, restricttoLKTs = NULL, list.data = NULL)
}
\description{
Makes MetagenomeSeq MRexperiments for every analysis that is possible to make given loaded jams files in list.data.
}
