% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SummarizedExperiments.R
\name{make_SummarizedExperiments}
\alias{make_SummarizedExperiments}
\title{make_SummarizedExperiments(pheno = NULL, onlysamples = NULL, onlyanalyses = c("LKT", "Product", "ECNumber", "Pfam", "Interpro", "resfinder", "PRINTS", "GO"), minnumsampanalysis = NULL, minpropsampanalysis = 0.1, stratify_functions_by_taxon = TRUE, add_TNF_data = FALSE, list.data = NULL, phenolabels = NULL, cdict = cdict, threads = 8, functional_contextualization_dissimilarity_cutoff = 0.15)}
\usage{
make_SummarizedExperiments(
  pheno = NULL,
  onlysamples = NULL,
  onlyanalyses = c("LKT", "Product", "ECNumber", "Pfam", "Interpro", "resfinder",
    "PRINTS", "GO"),
  minnumsampanalysis = NULL,
  minpropsampanalysis = 0.1,
  stratify_functions_by_taxon = TRUE,
  add_TNF_data = FALSE,
  list.data = NULL,
  phenolabels = NULL,
  cdict = cdict,
  threads = 8,
  functional_contextualization_dissimilarity_cutoff = 0.15
)
}
\description{
Makes a SummarizedExperiment object for every analysis that is possible to make given loaded jams files in list.data.
}
