% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_abundance.R
\name{pareto_abundance}
\alias{pareto_abundance}
\title{pareto_abundance(LKTdose = NULL, max_num_taxa_to_show = 30, orderby = "Percentage", readstats = NULL, read_denominator = "Assembled", samplename = NULL, taxlevel = NULL, showrelabundtext = TRUE, showcompletenesstext = TRUE, showcompletenessline = TRUE, showcontaminationline = FALSE, showparetocurve = TRUE, rotate_plot = FALSE)}
\usage{
pareto_abundance(
  LKTdose = NULL,
  max_num_taxa_to_show = 30,
  orderby = "Percentage",
  readstats = NULL,
  read_denominator = "Assembled",
  samplename = NULL,
  taxlevel = NULL,
  showrelabundtext = TRUE,
  showcompletenesstext = TRUE,
  showcompletenessline = TRUE,
  showcontaminationline = FALSE,
  showparetocurve = TRUE,
  rotate_plot = FALSE
)
}
\description{
Plots a pareto chart of the relative abundance of a taxon.
}
