% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_graph.R
\name{plot_bar_graph}
\alias{plot_bar_graph}
\title{Creates bar graph of relative taxonomic abundance from SummarizedExperiment object}
\usage{
plot_bar_graph(
  mgseqobj = NULL,
  glomby = NULL,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  units = 100,
  threshold = 2,
  title = "",
  category = NULL,
  cat_pos = -2,
  cat_text_size = 3,
  border_color = "white",
  cdict = NULL,
  colors = NULL,
  grid = TRUE,
  ...
)
}
\arguments{
\item{mgseqobj}{a SummarizedExperiment object.}

\item{glomby}{string giving taxonomic level (from tax_table) to plot at.}

\item{samplesToKeep}{vector with samples to plot.}

\item{featuresToKeep}{vector with features to plot.}

\item{units}{unit to normalize data to -- default percentage (100).}

\item{threshold}{minimum value to include as separate entry (below is Misc_Low_Abundance).}

\item{title}{plot title}

\item{category}{colData category to use for labels}

\item{cat_pos}{position of category line (change if ugly in plot)}

\item{cat_text_size}{text size of category labels}

\item{border_color}{color of bar borders}

\item{cdict}{use JAMS colour dictionary}

\item{colors}{use named list of taxon colors}
}
\description{
Creates bar graph of relative taxonomic abundance from SummarizedExperiment object
}
\examples{
plot_bar_graph(mrexp1, glomby = "Phylum", category = "Group")
}
