% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_graph.R
\name{plot_bar_graph}
\alias{plot_bar_graph}
\title{Creates bar graph of relative taxonomic abundance from SummarizedExperiment object}
\usage{
plot_bar_graph(
  ExpObj = NULL,
  glomby = NULL,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  groupby = NULL,
  colourby = NULL,
  subsetby = NULL,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL,
  PPM_normalize_to_bases_sequenced = FALSE,
  ignoreunclassified = FALSE,
  total_units = 100,
  threshold = 2,
  cat_pos = -2,
  cat_text_size = 3,
  legend_text_size = 4,
  border_color = "white",
  cdict = NULL,
  feature_cdict = NULL,
  addtit = NULL,
  grid = TRUE,
  class_to_ignore = "N_A",
  ...
)
}
\arguments{
\item{ExpObj}{a SummarizedExperiment object.}

\item{glomby}{string giving taxonomic level (from tax_table) to plot at.}

\item{samplesToKeep}{vector with samples to plot.}

\item{featuresToKeep}{vector with features to plot.}

\item{groupby}{colData category to use for labels}

\item{total_units}{unit to normalize data to -- default percentage (100).}

\item{threshold}{minimum value to include as separate entry (below is Misc_Low_Abundance).}

\item{cat_pos}{position of category line (change if ugly in plot)}

\item{cat_text_size}{text size of category labels}

\item{legend_text_size}{text size of legend labels}

\item{border_color}{color of bar borders}

\item{cdict}{use JAMS colour dictionary}

\item{feature_cdict}{use named list of taxon colors}

\item{addtit}{optional string to add more text to title}
}
\description{
Creates bar graph of relative taxonomic abundance from SummarizedExperiment object
}
\examples{
plot_bar_graph(ExpObj = mrexp1, glomby = "Phylum", groupby = "Group")
}
