% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_heatmap.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{plot_correlation_heatmap(ExpObj = NULL, glomby = NULL, stattype = "spearman", subsetby = NULL, maxnumfeatallowed = 10000, minabscorrcoeff = NULL, ntopvar = NULL, featuresToKeep = NULL, only_allow_CSBs = FALSE, samplesToKeep = NULL, applyfilters = NULL, featcutoff = NULL, GenomeCompletenessCutoff = NULL, show_GenomeCompleteness_boxplot = TRUE, PPM_normalize_to_bases_sequenced = FALSE, showGram = TRUE, showphylum = TRUE, addtit = NULL, cdict = NULL, ignoreunclassified = TRUE, class_to_ignore = "N_A", returnstats = FALSE)}
\usage{
plot_correlation_heatmap(
  ExpObj = NULL,
  glomby = NULL,
  stattype = "spearman",
  subsetby = NULL,
  maxnumfeatallowed = 10000,
  minabscorrcoeff = NULL,
  ntopvar = NULL,
  featuresToKeep = NULL,
  only_allow_CSBs = FALSE,
  samplesToKeep = NULL,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  show_GenomeCompleteness_boxplot = TRUE,
  PPM_normalize_to_bases_sequenced = FALSE,
  normalization = "relabund",
  showGram = TRUE,
  showphylum = TRUE,
  addtit = NULL,
  cdict = NULL,
  ignoreunclassified = TRUE,
  class_to_ignore = "N_A",
  returnstats = FALSE
)
}
\description{
Plots correlation heatmaps annotated by the metadata or a correlelogram of features
}
