% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_heatmap.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{plot_correlation_heatmap(ExpObj = NULL, glomby = NULL, heatpalette = "smart", stattype = "spearman", cluster_rows = FALSE, subsetby = NULL, maxnumfeatallowed = 500, featmaxatleastPPM = 0, featcutoff = c(0, 0), applyfilters = NULL, featuresToKeep = NULL, samplesToKeep = NULL, genomecompleteness = NULL, list.data = NULL, showGram = FALSE, showphylum = FALSE, addtit = NULL, mgSeqnorm = FALSE, cdict = NULL, ignoreunclassified = TRUE, returnstats = FALSE, ...)}
\usage{
plot_correlation_heatmap(
  ExpObj = NULL,
  glomby = NULL,
  stattype = "spearman",
  subsetby = NULL,
  maxnumfeatallowed = 10000,
  minabscorrcoeff = NULL,
  ntopvar = NULL,
  featmaxatleastPPM = 0,
  featcutoff = c(0, 0),
  applyfilters = NULL,
  featuresToKeep = NULL,
  samplesToKeep = NULL,
  genomecompleteness = NULL,
  list.data = NULL,
  showGram = TRUE,
  showphylum = TRUE,
  addtit = NULL,
  mgSeqnorm = FALSE,
  cdict = NULL,
  ignoreunclassified = TRUE,
  class_to_ignore = NULL
)
}
\description{
Plots correlation heatmaps annotated by the metadata or a correlelogram of features
}
