% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_heatmap.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{plot_correlation_heatmap(ExpObj = NULL, glomby = NULL, stattype = "spearman", subsetby = NULL, maxnumfeatallowed = 10000, minabscorrcoeff = NULL, ntopvar = NULL, featuresToKeep = NULL, samplesToKeep = NULL, applyfilters = NULL, featcutoff = NULL, GenomeCompletenessCutoff = NULL, PctFromCtgscutoff = NULL, PPM_normalize_to_bases_sequenced = FALSE, showGram = TRUE, showphylum = TRUE, addtit = NULL, cdict = NULL, ignoreunclassified = TRUE, class_to_ignore = NULL)}
\usage{
plot_correlation_heatmap(
  ExpObj = NULL,
  glomby = NULL,
  stattype = "spearman",
  subsetby = NULL,
  maxnumfeatallowed = 10000,
  minabscorrcoeff = NULL,
  ntopvar = NULL,
  featuresToKeep = NULL,
  samplesToKeep = NULL,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL,
  PPM_normalize_to_bases_sequenced = FALSE,
  showGram = TRUE,
  showphylum = TRUE,
  addtit = NULL,
  cdict = NULL,
  ignoreunclassified = TRUE,
  class_to_ignore = NULL
)
}
\description{
Plots correlation heatmaps annotated by the metadata or a correlelogram of features
}
