% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ordination.R
\name{plot_Ordination}
\alias{plot_Ordination}
\title{plot_Ordination(mgseqobj = NULL, glomby = NULL, subsetby = NULL, samplesToKeep = NULL, featuresToKeep = NULL, ignoreunclassified = TRUE, mgSeqnorm = FALSE, featmaxatleastPPM = 0, featcutoff = c(0, 0), applyfilters = NULL,  algorithm = "PCA", colourby = NULL, shapeby = NULL, sizeby = NULL, pairby = NULL, dotsize = 2, dotborder = NULL, log2tran = TRUE, transp = TRUE, perplx = NULL, permanova = FALSE, ellipse = FALSE, plotcentroids = FALSE, plottit = NULL, plot3D = FALSE, theta = 130, phi = 60, cdict = NULL, grid = TRUE, forceaspectratio = NULL, threads = 8, class_to_ignore = "N_A", ...)}
\usage{
plot_Ordination(
  mgseqobj = NULL,
  glomby = NULL,
  subsetby = NULL,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  ignoreunclassified = TRUE,
  mgSeqnorm = FALSE,
  featmaxatleastPPM = 0,
  featcutoff = c(0, 0),
  applyfilters = NULL,
  algorithm = "PCA",
  colourby = NULL,
  shapeby = NULL,
  sizeby = NULL,
  pairby = NULL,
  dotsize = 2,
  dotborder = NULL,
  log2tran = TRUE,
  transp = TRUE,
  perplx = NULL,
  permanova = FALSE,
  ellipse = FALSE,
  plotcentroids = FALSE,
  plottit = NULL,
  plot3D = FALSE,
  theta = 130,
  phi = 60,
  cdict = NULL,
  grid = TRUE,
  forceaspectratio = NULL,
  threads = 8,
  class_to_ignore = "N_A",
  ...
)
}
\description{
Creates ordination plots based on PCA, tSNE or tUMAP
}
