% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ordination.R
\name{plot_Ordination}
\alias{plot_Ordination}
\title{plot_Ordination(ExpObj = NULL, glomby = NULL, subsetby = NULL, samplesToKeep = NULL, samplesToHighlight = NULL, featuresToKeep = NULL, ignoreunclassified = TRUE, applyfilters = NULL, featcutoff = NULL, GenomeCompletenessCutoff = NULL, PctFromCtgscutoff = NULL, asPPM = TRUE, PPM_normalize_to_bases_sequenced = FALSE, assay_for_matrix = "BaseCounts", algorithm = "PCoA", PCA_Components = c(1, 2), distmethod = "bray", compareby = NULL, colourby = NULL, colorby = NULL, shapeby = NULL, use_letters_as_shapes = FALSE, sizeby = NULL, connectby = NULL, connection_orderby = NULL, textby = NULL, ellipseby = NULL, dotsize = 2, dotborder = NULL, log2tran = TRUE, transp = TRUE, perplx = NULL, max_neighbors = 15, permanova = TRUE, plotcentroids = FALSE, highlight_centroids = TRUE, show_centroid_distances = FALSE, calculate_centroid_distances_in_all_dimensions = FALSE, addtit = NULL, cdict = NULL, grid = TRUE, forceaspectratio = NULL, threads = 8, return_coordinates_matrix = FALSE, permanova_permutations = 10000, class_to_ignore = "N_A", ...)}
\usage{
plot_Ordination(
  ExpObj = NULL,
  glomby = NULL,
  subsetby = NULL,
  samplesToKeep = NULL,
  samplesToHighlight = NULL,
  featuresToKeep = NULL,
  ignoreunclassified = TRUE,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL,
  asPPM = TRUE,
  PPM_normalize_to_bases_sequenced = FALSE,
  assay_for_matrix = "BaseCounts",
  algorithm = "PCA",
  PCA_Components = c(1, 2),
  distmethod = "bray",
  compareby = NULL,
  colourby = NULL,
  colorby = NULL,
  shapeby = NULL,
  use_letters_as_shapes = FALSE,
  sizeby = NULL,
  connectby = NULL,
  connection_orderby = NULL,
  textby = NULL,
  ellipseby = NULL,
  dotsize = 2,
  dotborder = NULL,
  log2tran = TRUE,
  transp = TRUE,
  perplx = NULL,
  max_neighbors = 15,
  permanova = TRUE,
  plotcentroids = FALSE,
  highlight_centroids = TRUE,
  show_centroid_distances = FALSE,
  calculate_centroid_distances_in_all_dimensions = FALSE,
  addtit = NULL,
  cdict = NULL,
  grid = TRUE,
  forceaspectratio = NULL,
  threads = 8,
  return_coordinates_matrix = FALSE,
  permanova_permutations = 10000,
  class_to_ignore = "N_A",
  ...
)
}
\description{
Creates ordination plots based on PCA, tSNE or tUMAP
}
