% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relabund_features.R
\name{plot_relabund_features}
\alias{plot_relabund_features}
\title{plot_relabund_features(ExpObj = NULL, glomby = NULL, samplesToKeep = NULL, featuresToKeep = NULL, aggregatefeatures = FALSE, subsetby = NULL, compareby = NULL, colourby = NULL, shapeby = NULL, invertbinaryorder = FALSE, hmasPA = FALSE, applyfilters = NULL, featcutoff = NULL, GenomeCompletenessCutoff = NULL, PctFromCtgscutoff = NULL, ntop = NULL, adjustpval = FALSE, showpval = TRUE, showl2fc = TRUE, maxl2fc = NULL, minl2fc = NULL, addtit = NULL, PPM_normalize_to_bases_sequenced = FALSE, uselog = TRUE, statsonlog = FALSE, cdict = NULL, maxnumplots = NULL, signiflabel = "p.format", max_pairwise_cats = 4, numthreads = 1, nperm = 99, ignoreunclassified = TRUE, class_to_ignore = "N_A", ...)}
\usage{
plot_relabund_features(
  ExpObj = NULL,
  glomby = NULL,
  samplesToKeep = NULL,
  featuresToKeep = NULL,
  aggregatefeatures = FALSE,
  subsetby = NULL,
  compareby = NULL,
  colourby = NULL,
  shapeby = NULL,
  invertbinaryorder = FALSE,
  hmasPA = FALSE,
  applyfilters = NULL,
  featcutoff = NULL,
  GenomeCompletenessCutoff = NULL,
  PctFromCtgscutoff = NULL,
  ntop = NULL,
  adjustpval = FALSE,
  showpval = TRUE,
  showl2fc = TRUE,
  maxl2fc = NULL,
  minl2fc = NULL,
  addtit = NULL,
  PPM_normalize_to_bases_sequenced = FALSE,
  uselog = TRUE,
  statsonlog = FALSE,
  cdict = NULL,
  maxnumplots = NULL,
  signiflabel = "p.format",
  max_pairwise_cats = 4,
  numthreads = 1,
  nperm = 99,
  ignoreunclassified = TRUE,
  class_to_ignore = "N_A",
  ...
)
}
\description{
Generates relative abundance plots per feature annotated by the metadata using as input a SummarizedExperiment object
}
