% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wordcloud_sample.R
\name{plot_wordcloud_sample}
\alias{plot_wordcloud_sample}
\title{plot_wordcloud_sample(opt=opt, elements=NULL)}
\usage{
plot_wordcloud_sample(opt = NULL, analysis = NULL,
  removeunclassifieds = TRUE)
}
\description{
JAMSalpha function
}
