% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{retrieve_features_by_taxa}
\alias{retrieve_features_by_taxa}
\title{retrieve_features_by_taxa(wantedfeatures = NULL, wantedsamples = NULL, allfeaturesbytaxa_matrix = NULL, allfeaturesbytaxa_index = NULL)}
\usage{
retrieve_features_by_taxa(
  wantedfeatures = NULL,
  wantedsamples = NULL,
  allfeaturesbytaxa_matrix = NULL,
  allfeaturesbytaxa_index = NULL
)
}
\description{
Returns a long form data frame of stratification by taxa of wanted functional features in wanted samples, given allfeaturesbytaxa_matrix and allfeaturesbytaxa_index metadata present in a JAMS SummarizedExperiment functional object.
}
